poedalkin.ru/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
python-slugify==6.1.2
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.12
